+++ src/test/kotlin/de/magynhard/crystal/parser/CrystalParserTest.kt
@@ -238,4 +238,8 @@ class CrystalParserTest : ParsingTestCase("", "cr", CrystalParserDefinition()) {
     fun testLibFunPointerParams() {
         doTest(true)
     }
+
+    fun testNamespaceAccess() {
+        doTest(true)
+    }
 }